requirements.txt
ibm_db
Flask
NewsApiClient
pickle
requests
